fix: remove basic variant, adapt page
