+++ packages/angular-test-app/src/app/app-routing.module.ts
@@ -151,8 +151,7 @@ import MenuCategory from '../preview-examples/menu-category';
 import MenuWithBottomTabs from '../preview-examples/menu-with-bottom-tabs';
 import Message from '../preview-examples/message';
 import MessageBar from '../preview-examples/message-bar';
-import MessageBarDOMSingleRemoval from '../preview-examples/message-bar-dom-single-removal';
-import MessageBarDOMMultipleRemoval from '../preview-examples/message-bar-dom-multiple-removal';
+import MessageBarMultiple from '../preview-examples/message-bar-multiple';
 import ModalByInstance from '../preview-examples/modal-by-instance';
 import ModalByInstanceContent from '../preview-examples/modal-by-instance-content';
 import ModalByTemplate from '../preview-examples/modal-by-template';
@@ -719,12 +718,8 @@ const routes: Routes = [
         component: MessageBar,
       },
       {
-        path: 'message-bar-dom-single-removal',
-        component: MessageBarDOMSingleRemoval,
-      },
-      {
-        path: 'message-bar-dom-multiple-removal',
-        component: MessageBarDOMMultipleRemoval,
+        path: 'message-bar-multiple',
+        component: MessageBarMultiple,
       },
       {
         path: 'modal-by-instance-content',
