+++ packages/angular-test-app/src/app/app.module.ts
@@ -159,8 +159,7 @@ import MenuCategory from '../preview-examples/menu-category';
 import MenuWithBottomTabs from '../preview-examples/menu-with-bottom-tabs';
 import Message from '../preview-examples/message';
 import MessageBar from '../preview-examples/message-bar';
-import MessageBarDOMSingleRemoval from '../preview-examples/message-bar-dom-single-removal';
-import MessageBarDOMMultipleRemoval from '../preview-examples/message-bar-dom-multiple-removal';
+import MessageBarMultiple from '../preview-examples/message-bar-multiple';
 import ModalByInstance from '../preview-examples/modal-by-instance';
 import ModalByInstanceContent from '../preview-examples/modal-by-instance-content';
 import ModalByTemplate from '../preview-examples/modal-by-template';
@@ -374,8 +373,7 @@ import WorkflowVertical from '../preview-examples/workflow-vertical';
     MenuWithBottomTabs,
     Message,
     MessageBar,
-    MessageBarDOMSingleRemoval,
-    MessageBarDOMMultipleRemoval,
+    MessageBarMultiple,
     ModalByInstanceContent,
     ModalByInstance,
     ModalByTemplate,
