+++ packages/angular-test-app/src/preview-examples/message-bar.ts
@@ -10,19 +10,20 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-example',
+  selector: 'app-message-bar',
   template: `
     <div class="message-bar">
-      <ix-message-bar>Message text</ix-message-bar>
-      <ix-message-bar type="warning">Message text</ix-message-bar>
-      <ix-message-bar type="danger">
-        <div class="d-flex align-items-center justify-content-between">
-          Message text
-          <ix-button>Action</ix-button>
-        </div>
+      <ix-message-bar *ngIf="showMessageBar" (closeAnimationCompleted)="handleCloseAnimationCompleted()">
+        Message text
       </ix-message-bar>
     </div>
   `,
-  styleUrls: ['./message-bar.css'],
+  styleUrls: ['./message-bar.css']
 })
-export default class MessageBar {}
+export default class MessageBar {
+  showMessageBar = true;
+
+  handleCloseAnimationCompleted() {
+    this.showMessageBar = false;
+  }
+}
