+++ packages/documentation/docs/controls/_message_bar_code.mdx
@@ -7,6 +7,9 @@ import Playground from '@site/src/components/PlaygroundV3';
 ### Examples
 
 #### Basic
+The message bar web component only provides the visual appearance of the message bar.
+To fully utilize the message bar, you need to implement a mechanism to remove it from the DOM when it is no longer needed.
+This typically involves handling the close event and updating the state of your application to reflect the removal of the message bar.
 
 <Playground
   name="message-bar"
@@ -14,23 +17,10 @@ import Playground from '@site/src/components/PlaygroundV3';
 >
 </Playground>
 
-#### Removal
-The message bar web component only provides the visual appearance of the message bar.
-To fully utilize the message bar, you need to implement a mechanism to remove it from the DOM when it is no longer needed.
-This typically involves handling the close event and updating the state of your application to reflect the removal of the message bar.
-
-#### Single
-
-<Playground
-  name="message-bar-dom-single-removal"
-  height="14rem"
->
-</Playground>
-
 #### Multiple
 
 <Playground
-  name="message-bar-dom-multiple-removal"
+  name="message-bar-multiple"
   height="14rem"
 >
 </Playground>
