--- packages/html-test-app/src/preview-examples/message-bar-dom-multiple-removal.html
@@ -1,41 +1,0 @@
-<!--
-SPDX-FileCopyrightText: 2024 Siemens AG
-
-SPDX-License-Identifier: MIT
-
-This source code is licensed under the MIT license found in the
-LICENSE file in the root directory of this source tree.
--->
-
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Message bar example</title>
-    <link rel="stylesheet" href="./message-bar.css" />
-  </head>
-
-  <body>
-    <ix-button id="showMessageBarButton">Show Message Bar</ix-button>
-    <div class="message-bar" id="messageBarContainer"></div>
-
-    <script type="module">
-      const variants = ['info', 'warning', 'danger'];
-
-      document.getElementById('showMessageBarButton').addEventListener('click', () => {
-        const messageBarContainer = document.getElementById('messageBarContainer');
-        const randomVariant = variants[Math.floor(Math.random() * variants.length)];
-        const messageBar = document.createElement('ix-message-bar');
-        messageBar.setAttribute('type', randomVariant);
-        messageBar.textContent = 'Message text';
-        messageBar.addEventListener('closeAnimationCompleted', () => {
-          messageBarContainer.removeChild(messageBar);
-        });
-        messageBarContainer.appendChild(messageBar);
-      });
-    </script>
-
-    <script type="module" src="./init.js"></script>
-  </body>
-</html>
