+++ packages/html-test-app/src/preview-examples/message-bar-multiple.html
@@ -1,0 +1,49 @@
+<!--
+SPDX-FileCopyrightText: 2024 Siemens AG
+
+SPDX-License-Identifier: MIT
+
+This source code is licensed under the MIT license found in the
+LICENSE file in the root directory of this source tree.
+-->
+
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Message bar example</title>
+  <link rel="stylesheet" href="./message-bar.css" />
+</head>
+
+<body>
+<div class="message-bar" id="messageBarContainer"></div>
+
+<script type="module">
+  const messageBarContainer = document.getElementById('messageBarContainer');
+
+  const createMessageBar = (variant) => {
+    const messageBar = document.createElement('ix-message-bar');
+    messageBar.textContent = 'Message text';
+    if (variant) {
+      messageBar.setAttribute('type', variant);
+    }
+    messageBar.addEventListener('closeAnimationCompleted', () => {
+      messageBarContainer.removeChild(messageBar);
+    });
+    return messageBar;
+  };
+
+  const messageBars = [
+    createMessageBar('info'),
+    createMessageBar('warning'),
+    createMessageBar('danger')
+  ];
+
+  messageBars.forEach(messageBar => {
+    messageBarContainer.appendChild(messageBar);
+  });
+</script>
+<script type="module" src="./init.js"></script>
+</body>
+</html>
